coordinator: extract inject_test_packet into coordinator/inject.rs
#985 Phase 8 — second impl-block extraction. Move the
`inject_test_packet` RPC handler (~140 LOC) out of coordinator/mod.rs
into a sibling `impl super::Coordinator` block.

inject_test_packet builds a synthetic packet against the live
ForwardingState/HA snapshot, runs it through the resolution path,
and reports the disposition. It mutates Coordinator only to fill
`last_resolution`; otherwise lifecycle-free. Cohesive RPC handler
that doesn't belong in mod.rs's lifecycle-management body.

coordinator/mod.rs production LOC (excluding mod tests):
  2,028 → 1,888 — first time UNDER the 2,000 modularity-discipline
  threshold since this refactor stream started.

inject.rs: 137 LOC.

Per docs/refactor/985-coordinator-decomp.md (origin/research/985-
coordinator-design). Phase 7 (status accessors) in #1031; Phases
0-6 are field-extractions (#1019/#1023/#1024/#1025/#1027/#1029/#1030).

build clean, 865 Rust tests pass.
